requirements.txt
fastapi
uvicorn
python-multipart
line-bot-sdk
tensorflow
numpy
Pillow
ml_model
predictor